Update 1: fix bugs

+++ utils/processing.py
@@ -6,7 +6,7 @@ import torch
 
 class CustomDataset(torch.utils.data.Dataset):
     def __init__(self, input_data, targets, transform=None):
-        self.data = [input_data[i].unsqueeze(0) for i in range(input_data.size(0))]
+        self.data = [input_data[i].unsqueeze(0).float() for i in range(input_data.size(0))]
         self.targets = targets
         self.classes = torch.unique(targets).tolist()
 
@@ -15,6 +15,7 @@ class CustomDataset(torch.utils.data.Dataset):
 
     def __getitem__(self, idx):
         return self.data[idx], self.targets[idx]
+
 
 def set_seed(seed_value):
     random.seed(seed_value)
